% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fevd.R
\name{fevd_plot}
\alias{fevd_plot}
\title{Plot all FEVDs}
\usage{
fevd_plot(fevd, responses = NULL, verticle = FALSE)
}
\arguments{
\item{fevd}{fevd object}

\item{responses}{string vector: responses to plot}

\item{verticle}{boolean: If true then stack all plots into one column}
}
\value{
grid of ggplot2 graphs
}
\description{
Plot all FEVDs
}
