% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecasts.R
\name{forecast_plot}
\alias{forecast_plot}
\title{Chart  forecast}
\usage{
forecast_plot(forecasts, series = NULL, verticle = FALSE)
}
\arguments{
\item{forecasts}{data.frame: sovereign forecast object}

\item{series}{string: series to plot (default to all series)}

\item{verticle}{boolean: If true then stack all plots into one column}
}
\value{
grid of ggplot2 graphs
}
\description{
Chart  forecast
}
