% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fevd.R
\name{individual_fevd_plot}
\alias{individual_fevd_plot}
\title{Plot an individual IRF}
\usage{
individual_fevd_plot(fevd, response.var, title, ylab)
}
\arguments{
\item{fevd}{fevd object}

\item{response.var}{string: name of variable to treat as the response}

\item{title}{string: title of the chart}

\item{ylab}{string: y-axis label}
}
\value{
ggplot2 graph
}
\description{
Plot an individual IRF
}
