% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_irf.R
\name{individual_lp_irf_plot}
\alias{individual_lp_irf_plot}
\title{Plot an individual IRF}
\usage{
individual_lp_irf_plot(irf, shock.var, response.var, title, ylab)
}
\arguments{
\item{shock.var}{string: name of variable to treat as the shock}

\item{response.var}{string: name of variable to treat as the response}

\item{title}{string: title of the chart}

\item{ylab}{string: y-axis label}

\item{irfs}{var_irf object}
}
\value{
ggplot2 graph
}
\description{
Plot an individual IRF
}
