% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimes.R
\name{learn_regimes}
\alias{learn_regimes}
\title{Assign regimes via unsupervised machine learning methods}
\usage{
learn_regimes(data, regime.n = 2, engine = "rf")
}
\arguments{
\item{data}{data.frame, matrix, ts, xts, zoo: Endogenous regressors}

\item{regime.n}{int: number of regimes to estimate (only applies to kmeans)}

\item{engine}{string: regime assignment technique ('rf' or 'kmeans')}
}
\value{
\code{data} as a data.frame with a regime column assigning rows to mutually exclusive regimes. If engine = 'rf' is used then regime probabilities will be returned as well.
}
\description{
Assign regimes via unsupervised machine learning methods
}
\examples{
\dontrun{

  learn_regimes(
     data = Data,
     regime.n = 3,
     engine = 'kmeans')
}

}
