% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{LP}
\alias{LP}
\title{Estimate local projections}
\usage{
LP(data, p = 1, horizon = 1, freq = "month")
}
\arguments{
\item{data}{data.frame, matrix, ts, xts, zoo: Endogenous regressors}

\item{p}{int: lags}

\item{horizon}{int: forecast horizons}

\item{freq}{string: frequency of data (day, week, month, quarter, year)}
}
\value{
list object with elements \code{data}, \code{model}, \code{forecasts}, \code{residuals}; if there is more than one forecast horizon estimated, then \code{model}, \code{forecasts}, \code{residuals} will each be a list where each element corresponds to a single horizon
}
\description{
Estimate local projections
}
\examples{
\dontrun{
LP(
  data = Data,
  p = 1,
  horizon = 1,
  freq = 'month')
}

}
