% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_irf.R
\name{lp_irf}
\alias{lp_irf}
\title{Estimate single-regime impulse response functions}
\usage{
lp_irf(lp, CI = c(0.1, 0.9))
}
\arguments{
\item{lp}{LP output}

\item{CI}{numeric vector: c(lower ci bound, upper ci bound)}
}
\value{
long-form data.frame with one row per target-shock-horizon identifier
}
\description{
Estimate single-regime impulse response functions
}
\examples{
\dontrun{
lp =
  LP(
    data = Data,
    p = 1,
    horizon = 1,
    freq = 'month')

irf =
  lp_irf(
    lp,
    CI = c(0.05,0.95))
}

}
