% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_irf_chart.R
\name{lp_irf_chart}
\alias{lp_irf_chart}
\title{Chart local projection IRFs}
\usage{
lp_irf_chart(plotData, title, ylim, ystep, ylab)
}
\arguments{
\item{plotData}{lp_irf output}

\item{title}{string: chart title}

\item{ylim}{int: y-axis limits, c(lower limit, upper limit)}

\item{ystep}{int: step size inbetween y-axis tick marks}

\item{ylab}{string: y-axis label}
}
\value{
plot
}
\description{
Chart local projection IRFs
}
