% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_irf.R
\name{lp_irf_plot}
\alias{lp_irf_plot}
\title{Plot all IRFs}
\usage{
lp_irf_plot(irfs, shocks = NULL, responses = NULL, verticle = FALSE)
}
\arguments{
\item{irfs}{lp_irf object}

\item{shocks}{string vector: shocks to plot}

\item{responses}{string vector: responses to plot}
}
\value{
grid of ggplot2 graphs
}
\description{
Plot all IRFs
}
