% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fevd.R
\name{plot_fevd}
\alias{plot_fevd}
\title{Chart FEVDs}
\usage{
plot_fevd(fevd, responses = NULL, verticle = FALSE)
}
\arguments{
\item{fevd}{fevd object}

\item{responses}{string vector: responses to plot}

\item{verticle}{boolean: If true then stack all plots into one column}
}
\value{
grid of ggplot2 graphs
}
\description{
Chart FEVDs
}
