% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hd.R
\name{plot_hd}
\alias{plot_hd}
\title{Chart HDs}
\usage{
plot_hd(hd, verticle = FALSE)
}
\arguments{
\item{hd}{hd object}

\item{verticle}{boolean: If true then stack all plots into one column}
}
\value{
grid of ggplot2 graphs
}
\description{
Chart HDs
}
