% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fevd.R
\name{plot_individual_fevd}
\alias{plot_individual_fevd}
\title{Plot an individual FEVD}
\usage{
plot_individual_fevd(fevd, response.var, title, ylab)
}
\arguments{
\item{fevd}{fevd object}

\item{response.var}{string: name of variable to treat as the response}

\item{title}{string: title of the chart}

\item{ylab}{string: y-axis label}
}
\value{
ggplot2 graph
}
\description{
Plot an individual FEVD
}
