% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hd.R
\name{plot_individual_hd}
\alias{plot_individual_hd}
\title{Plot an individual HD}
\usage{
plot_individual_hd(hd, target.var, title)
}
\arguments{
\item{hd}{hd object}

\item{target.var}{string: name of variable to decompose into shocks}

\item{title}{string: title of the chart}
}
\value{
ggplot2 graph
}
\description{
Plot an individual HD
}
