% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_irf.R
\name{plot_irf}
\alias{plot_irf}
\title{Chart IRFs}
\usage{
plot_irf(irf, shocks = NULL, responses = NULL, verticle = FALSE)
}
\arguments{
\item{irf}{irf object}

\item{shocks}{string vector: shocks to plot}

\item{responses}{string vector: responses to plot}

\item{verticle}{boolean: If true then stack all plots into one column}
}
\value{
grid of ggplot2 graphs
}
\description{
Chart IRFs
}
