% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{threshold_LP}
\alias{threshold_LP}
\title{Estimate threshold local projections}
\usage{
threshold_LP(
  data,
  regime,
  p = 1,
  horizons = 1,
  freq = "month",
  NW = FALSE,
  NW_lags = NULL,
  NW_prewhite = TRUE
)
}
\arguments{
\item{data}{data.frame, matrix, ts, xts, zoo: Endogenous regressors}

\item{regime}{string: name or regime assignment vector in the design matrix (data)}

\item{p}{int: lags}

\item{horizons}{int: forecast horizons}

\item{freq}{string: frequency of data (day, week, month, quarter, year)}
}
\value{
list object with elements \code{data}, \code{model}, \code{forecasts}, \code{residuals}; if there is more than one forecast horizon estimated, then \code{model}, \code{forecasts}, \code{residuals} will each be a list where each element corresponds to a single horizon
}
\description{
Estimate threshold local projections
}
\examples{
\dontrun{
threshold_LP(
  data = Data,
  regime = 'regime',
  p = 1,
  horizon = 1,
  freq = 'month')
}

}
