% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_projections.R
\name{threshold_lp_irf}
\alias{threshold_lp_irf}
\title{Estimate multi-regime local projection IRFs}
\usage{
threshold_lp_irf(data, shock, target, thresholdVar = NULL, horizons, lags)
}
\arguments{
\item{data}{data.frame, matrix, ts, xts, zoo: Endogenous regressors}

\item{shock}{string: variable to shock}

\item{target}{string: variable betas to collect}

\item{thresholdVar}{data.frame of regime binaries}

\item{horizons}{int: horizons to forecast out to}

\item{lags}{int: lags to include in regressions}
}
\value{
data.frame
}
\description{
Estimate multi-regime local projection IRFs
}
\examples{
\dontrun{
threshold_lp_irf(
  data = Data,
  shock = 'x',
  target = 'y',
  thresholdVar = Data.regime,
  horizons = 20,
  lags = 2)
}

}
