% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{threshold_VAR}
\alias{threshold_VAR}
\title{Estimate multi-regime VAR}
\usage{
threshold_VAR(data, regime, p = 1, horizon = 10, freq = "month")
}
\arguments{
\item{data}{data.frame, matrix, ts, xts, zoo: Endogenous regressors}

\item{regime}{string: name or regime assignment vector in the design matrix (data)}

\item{p}{int: lags}

\item{horizon}{int: forecast horizons}

\item{freq}{string: frequency of data (day, week, month, quarter, year)}
}
\value{
list of lists, each regime returns its own list with elements \code{data}, \code{model}, \code{forecasts}, \code{residuals}
}
\description{
Estimate multi-regime VAR
}
\examples{
\dontrun{
threshold_VAR(
  data = Data,
  regime = 'regime',
  p = 1,
  horizon = 10,
  freq = 'month')
}

}
