% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_fevd.R
\name{threshold_var_fevd}
\alias{threshold_var_fevd}
\title{Estimate multi-regime forecast error variance decomposition}
\usage{
threshold_var_fevd(threshold_var, horizon = 10, scale = TRUE)
}
\arguments{
\item{threshold_var}{threshold_var output}

\item{horizon}{int: number of periods}

\item{scale}{boolean: scale variable contribution as percent of total error}
}
\value{
list, each regime returns its own long-form data.frame
}
\description{
Estimate multi-regime forecast error variance decomposition
}
\examples{
\dontrun{
threshold_var_fevd(
  threshold_var,
  bootstraps.num = 10)
}

}
