% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_irf.R
\name{var_irf_plot}
\alias{var_irf_plot}
\title{Plot all IRFs}
\usage{
var_irf_plot(irfs, shocks = NULL, responses = NULL)
}
\arguments{
\item{irfs}{var_irf object}

\item{shocks}{string vector: shocks to plot}

\item{responses}{string vector: responses to plot}
}
\value{
grid of ggplot2 graphs
}
\description{
Plot all IRFs
}
