% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{LP}
\alias{LP}
\title{Estimate single-regime local projections}
\usage{
LP(
  data,
  p = 1,
  horizons = 1,
  freq = "month",
  NW = FALSE,
  NW_lags = NULL,
  NW_prewhite = TRUE
)
}
\arguments{
\item{data}{data.frame, matrix, ts, xts, zoo: Endogenous regressors}

\item{p}{int: lags}

\item{horizons}{int: forecast horizons}

\item{freq}{string: frequency of data (day, week, month, quarter, year)}

\item{NW}{boolean: Newey-West correction on variance-covariance matrix}

\item{NW_lags}{int: number of lags to use in Newey-West correction}

\item{NW_prewhite}{boolean: TRUE prewhite option for Newey-West correction (see sandwich::NeweyWest function)}
}
\value{
list object with elements \code{data}, \code{model}, \code{forecasts}, \code{residuals}; if there is more than one forecast horizon estimated, then \code{model}, \code{forecasts}, \code{residuals} will each be a list where each element corresponds to a single horizon
}
\description{
Estimate single-regime local projections
}
\examples{
\dontrun{
LP(
  data = Data,
  p = 1,
  horizon = 1,
  freq = 'month')
}

}
