% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_projections.R
\name{lp_irf}
\alias{lp_irf}
\title{Estimate single-regime local projection IRFs}
\usage{
lp_irf(data, shock, target, horizons, lags)
}
\arguments{
\item{data}{data.frame, matrix, ts, xts, zoo: Endogenous regressors}

\item{shock}{string: variable to shock}

\item{target}{string: variable betas to collect}

\item{horizons}{int: horizons to forecast out to}

\item{lags}{int: lags to include in regressions}
}
\value{
data.frame
}
\description{
Estimate single-regime local projection IRFs
}
\examples{
\dontrun{
lp_irf(
  data = Data,
  shock = 'x',
  target = 'y',
  horizons = 20,
  lags = 2)
}

}
