% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_irf.R
\name{rvar_irf}
\alias{rvar_irf}
\title{Estimate multi-regime impulse response functions}
\usage{
rvar_irf(rvar, horizon = 10, bootstraps.num = 100, CI = c(0.1, 0.9))
}
\arguments{
\item{rvar}{rvar output}

\item{horizon}{int: number of periods}

\item{bootstraps.num}{int: number of bootstraps}

\item{CI}{numeric vector: c(lower ci bound, upper ci bound)}
}
\value{
list of lists, each regime returns its own list with elements \code{irfs}, \code{ci.lower}, and \code{ci.upper}; all elements are long-form data.frames
}
\description{
Estimate multi-regime impulse response functions
}
\examples{
\dontrun{
rvar_irf(
  rvar,
  bootstraps.num = 10,
  CI = c(0.05,0.95))
}

}
