% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_irf.R
\name{threshold_lp_irf}
\alias{threshold_lp_irf}
\title{Estimate multi-regime impulse response functions}
\usage{
threshold_lp_irf(threshold_lp, CI = c(0.1, 0.9))
}
\arguments{
\item{threshold_lp}{threshold_LP output}

\item{CI}{numeric vector: c(lower ci bound, upper ci bound)}
}
\value{
list of long-form data.frame with one row per target-shock-horizon identifier
}
\description{
Estimate multi-regime impulse response functions
}
\examples{
\dontrun{
tlp =
  threshold_LP(
    data = Data,
    regime = 'regime',
    p = 1,
    horizon = 1,
    freq = 'month')

tirf =
  threshold_lp_irf(
    threshold_lp = tlp,
    CI = c(0.05,0.95))
}

}
