% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{VAR}
\alias{VAR}
\title{Estimate single-regime VAR}
\usage{
VAR(data, p = 1, horizon = 10, freq = "month")
}
\arguments{
\item{data}{data.frame, matrix, ts, xts, zoo: Endogenous regressors}

\item{p}{int: lags}

\item{horizon}{int: forecast horizons}

\item{freq}{string: frequency of data (day, week, month, quarter, year)}
}
\value{
list object with elements \code{data}, \code{model}, \code{forecasts}, \code{residuals}
}
\description{
Estimate single-regime VAR
}
\examples{
\dontrun{
VAR(
  data = Data,
  p = 1,
  horizon = 10,
  freq = 'month')
}

}
