% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_fevd.R
\name{var_fevd}
\alias{var_fevd}
\title{Estimate single-regime forecast error variance decomposition}
\usage{
var_fevd(var, horizon = 10, scale = TRUE)
}
\arguments{
\item{var}{VAR output}

\item{horizon}{int: number of periods}

\item{scale}{boolean: scale variable contribution as percent of total error}
}
\value{
long-form data.frame
}
\description{
Estimate single-regime forecast error variance decomposition
}
\examples{
\dontrun{
var_fevd(
  var,
  bootstraps.num = 10)
}

}
