% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_irf.R
\name{var_irf}
\alias{var_irf}
\title{Estimate single-regime impulse response functions}
\usage{
var_irf(var, horizon = 10, bootstraps.num = 100, CI = c(0.1, 0.9))
}
\arguments{
\item{var}{VAR output}

\item{horizon}{int: number of periods}

\item{bootstraps.num}{int: number of bootstraps}

\item{CI}{numeric vector: c(lower ci bound, upper ci bound)}
}
\value{
list object with elements \code{irfs}, \code{ci.lower}, and \code{ci.upper}; all elements are long-form data.frames
}
\description{
Estimate single-regime impulse response functions
}
\examples{
\dontrun{
var_irf(
  var,
  bootstraps.num = 10,
  CI = c(0.05,0.95))
}

}
